requirements.txt
Django==2.0.4
django-cors-headers==2.2.0
django-model-utils==3.1.1
djangorestframework==3.7.7
drf-nested-routers==0.90.2
psycopg2==2.7.4
psycopg2-binary==2.7.4
pytz==2018.4
